+++ feature_book_details/src/main/java/com/allsoftdroid/feature/book_details/utils/NetworkState.kt
@@ -1,4 +1,4 @@
-package com.allsoftdroid.feature.book_details.presentation
+package com.allsoftdroid.feature.book_details.utils
 
 enum class NetworkState {
     ERROR,
